% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootrstrappProcess.R
\name{bootstrapProcess}
\alias{bootstrapProcess}
\title{Show animated/still bootstapping process}
\usage{
bootstrapProcess(s, anime)
}
\arguments{
\item{s}{sample dataset in a vector form}

\item{anime}{boolean, where TRUE = T = 1, FALSE = F = 0}
}
\value{
A time series cumulative animation.
}
\description{
Takes in dataset in a vector form. If anime = TRUE, returns animated time series plot which shows elements sampled with replacement, otherwise returns still plot of bootstrap sample.
}
\examples{
bootstrapProcess(rnorm(30), TRUE)

}
\author{
Gulzina Kuttubekova
}
