% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapSample.R
\name{bootstrapSample}
\alias{bootstrapSample}
\title{Generate bootstrap samples}
\usage{
bootstrapSample(s, n)
}
\arguments{
\item{s}{sample dataset in a vector form}

\item{n}{number of bootstrap samples to generate}
}
\value{
A list of n bootstrap samples
}
\description{
Takes in dataset in a vector form. Returns a list of n bootstrap sample.
}
\author{
Gulzina Kuttubekova
}
