% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypTest_prompt.R
\name{hypTest_prompt}
\alias{hypTest_prompt}
\title{Returns a randomly selected hypothesis test prompt}
\usage{
hypTest_prompt()
}
\value{
a randomly selected hypothesis test prompt as a list
}
\description{
Returns a randomly selected hypothesis test prompt as a list
}
\examples{
hypTest_prompt()

}
\author{
Joe Zemmels
}
