% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errors.R
<<<<<<< HEAD:man/Errors.Rd
\name{Errors}
\alias{Errors}
\alias{errors}
\title{Plot Type I and Type II Error}
\usage{
errors(means = 0, sds = 1, alpha = 0.05,
=======
\name{ploterrors}
\alias{ploterrors}
\title{Plot Type I and Type II Error}
\usage{
ploterrors(means = 0, sds = 1, alpha = 0.05,
>>>>>>> e064be24d63e689506f34ff93a163d89ba0c6555:man/ploterrors.Rd
  direction = intToUtf8("8800"), alt = 1, plotly = FALSE)
}
\arguments{
\item{means}{mean of normal distribution}

\item{sds}{standard deviation of normal distribution}

\item{alpha}{significance level}

\item{direction}{direction of alternative hypothesis}

\item{plotly}{specifies whether a plotly object is desired}
}
\value{
A ggplot or plotly object, depending on whether the plotly argument was specified TRUE or FALSE.
}
\description{
Returns a ggplot or plotly object of a normal distribution with a specified significance level
}
\author{
Eryn Blagg
}
