% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModal.R
\name{plotModal}
\alias{plotModal}
\title{Plot a random distribution with differing modality}
\usage{
plotModal(randomDist = 1)
}
\arguments{
\item{randomDist}{Randomly generated number to signify type of distribution}
}
\value{
A list consisting of one character string and a ggplot object
}
\description{
Returns a ggplot object of certain modality (unimodal, bimodal, uniform)
}
\author{
Charlotte Roiger
}
