% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNormal.R
\name{plotNormal}
\alias{plotNormal}
\title{Plot a normal distribution}
\usage{
plotNormal(mu = 0, sigma = 1, alpha = 0.05, obs = NULL,
  direction = intToUtf8("8800"), plotly = FALSE)
}
\arguments{
\item{mu}{mean of normal distribution}

\item{sigma}{standard deviation of normal distribution}

\item{alpha}{significance level}

\item{obs}{observed sample mean (random generated if left blank)}

\item{direction}{direction of alternative hypothesis}

\item{plotly}{specifies whether a plotly object is desired}
}
\value{
A ggplot or plotly object, depending on whether the plotly argument was specified TRUE or FALSE.
}
\description{
Returns a ggplot or plotly object of a normal distribution with a specified significance level
}
\examples{
plotNormal() #default standard normal ggplot
plotNormal(mu=5,sigma=2,alpha=.1,obs=.5,plotly=TRUE) #customized plotly

}
\author{
Joe Zemmels
}
