% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutlier.R
\name{plotOutlier}
\alias{plotOutlier}
\title{Plot a random distribution with differing outliers}
\usage{
plotOutlier(randomDist = 1)
}
\arguments{
\item{randomDist}{Randomly generated number to signify type of distribution}
}
\value{
A list consisting of one character string and a ggplot object
}
\description{
Returns a ggplot object of a certain type of outlier (none, one, multiple)
}
\author{
Charlotte Roiger
}
