% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShape.R
\name{plotShape}
\alias{plotShape}
\title{Plot a random distribution with differing shape}
\usage{
plotShape(randomDist = 1)
}
\arguments{
\item{randomDist}{Randomly generated number to signify type of distribution}
}
\value{
A list consisting of one character string and a ggplot object
}
\description{
Returns a ggplot object of certain shape (right-skewed, left-skewed, symmetric)
}
\author{
Charlotte Roiger
}
