% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSamplingDistribution.R
\name{randomSample_histogram}
\alias{randomSample_histogram}
\title{Generate a histogram of a random sample}
\usage{
randomSample_histogram(sampleData, binwidth = 1,
  variableName = "Height", plotly = FALSE)
}
\arguments{
\item{sampleData}{data frame of random samples from a normal distribution}

\item{binwidth}{width of bins in histogram}

\item{variableName}{name of variable to be labeled on the x-axis of histogram}

\item{plotly}{specifies whether a plotly object is desired}
}
\value{
A ggplot or plotly object, depending on whether the plotly argument was specified TRUE or FALSE.
}
\description{
Returns a ggplot or plotly object of a histogram generated from a random sample
}
\examples{
dat <- randomSample(mu=1,sigma=4,sampleSize=100,numSamples=5)
randomSample_histogram(dat)

}
\author{
Joe Zemmels
}
