% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Launch the statsfoRstudents shiny app}
\source{
https://deanattali.com/2015/04/21/r-package-shiny-app/
}
\usage{
runShiny()
}
\description{
Launches the statsfoRstudents shiny app in the active R session
}
\examples{
\dontrun{
runShiny()
}
}
\author{
Joe Zemmels
}
