% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Launch a Shiny App locally}
\source{
https://deanattali.com/2015/04/21/r-package-shiny-app/
}
\usage{
runShinyApp(example)
}
\arguments{
\item{example}{name of shiny app to be launched. Current options are: "joe"}
}
\description{
Launch a Shiny App from the inst/shiny-examples folder in this package
}
\examples{
\dontrun{
runShinyApp() #to see a list of launchable apps in the inst/shiny-examples folder
runShinyApp("joe") #to launch Joe's shiny app
}
}
\author{
Joe Zemmels
}
