% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSamplingDistribution.R
\name{updateSampleMeans}
\alias{updateSampleMeans}
\title{Create/update a data frame of sample means}
\usage{
updateSampleMeans(sampleData, sampleMeans = NULL)
}
\arguments{
\item{sampleMeans}{data frame of sample means (empty data frame is initialized if this is argument is blank)}

\item{randomSample}{data frame of random samples from a normal distribution}
}
\value{
A data frame of sample means from each sample in randomSample
}
\description{
Creates or updates a data frame of sample means from a randomSample object passed to it
}
\examples{
dat <- randomSample(numSamples=100)
head(updateSampleMeans(sampleData=dat))

}
\author{
Joe Zemmels
}
