% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Launch the statsfoRstudents shiny app}
\usage{
runShiny()
}
\description{
Launches the statsfoRstudents shiny app in the active R session
}
\examples{
runShiny()

}
\author{
Joe Zemmels
}
