% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData_means.R
\docType{data}
\name{sampleData_means}
\alias{sampleData_means}
\title{Sample means of 100 normal random samples}
\format{A data frame of one column with 100 rows:
\describe{
\item{means}{mean of each sample}
}}
\source{
\url{https://github.com/jzemmels/statsfoRdummies/blob/master/R/meanSamplingDistribution.R}
}
\usage{
sampleData_means
}
\description{
A dataset containing the sample means of 100 normal random samples of population mean 5, population standard deviation 2, and sample size 25. Can be plotted using the sampleMeans_histogram() function.
}
\keyword{datasets}
